+++ typescript/src/lib/extractors/jira/extractor.ts
@@ -1,4 +1,3 @@
-import { toCSV, toSerializedArray } from './components/exporter';
 import { extractBatchFromConfig, extractAllFromConfig } from './components/extract';
 import { importConfig } from './components/import-config';
 import { IWorkItem } from '../../core/types';
@@ -7,6 +6,7 @@ import { IJiraSettings} from './types';
 
 class JiraExtractor {
   config: IJiraSettings = null;
+
   batchSize: number;
 
   constructor(config?: IJiraSettings) {
@@ -46,14 +46,26 @@ class JiraExtractor {
   toSerializedArray(workItems, withHeader?) {
     return toSerializedArray(workItems, Object.keys(this.config.Workflow), this.config.Attributes, withHeader);
   };
-
-  // testConnection = testConnection;
-  // getProjects = getProjects;
-  // getWorkflows = getWorkflows;
-  // getIssues = getIssues;
-  // getMetadata = getMetadata;
 };
 
 export {
   JiraExtractor,
+};
+
+
+const toCSV = (workItems: IWorkItem[], stages: string[], attributes: {}, domainUrl: string, withHeader: boolean = true, config: any = {}): string => {
+  if (attributes === undefined || attributes === null) {
+    attributes = {};
+  }
+  const header = `ID,Link,Name,${stages.join(',')},Type,${Object.keys(attributes).join(',')}`;
+  const body = workItems.map(item => item.toCSV(domainUrl, config)).reduce((res, cur) => `${res + cur}\n`, '');
+  const csv: string = `${header}\n${body}`;
+  return csv;
+};
+
+const toSerializedArray = (workItems: IWorkItem[], stages: string[], attributes: {}, withHeader: boolean = true): string => {
+  const header = `["ID","Link","Name",${stages.map(stage => `"${stage}"`).join(',')},"Type",${Object.keys(attributes).map(attribute => `"${attribute}"`).join(',')}]`;
+  const body = workItems.map(item => item.toSerializedArray()).reduce((res, cur) => `${res},\n${cur}`, '');
+  const serializedData: string = `[${header}${body}]`;
+  return serializedData;
 };
